Add E2E test for NoCode Mobile local device
