+++ e2e-test/mobile/delay.ts
@@ -1,0 +1,4 @@
+export const delay = (ms: number) =>
+  new Promise((resolve) => {
+    setTimeout(resolve, ms);
+  });
