+++ e2e-test/mobile/kill-start-process.ts
@@ -1,0 +1,5 @@
+import { spawn } from "node:child_process";
+
+export const killStartProcess = () => {
+  spawn("pkill", ["-f", "run link start"]);
+};
