+++ package.json
@@ -57,12 +57,14 @@
     "eslint-config-oclif": "^6.0.32",
     "eslint-config-prettier": "^10.1.1",
     "eslint-plugin-unicorn": "^57.0.0",
-    "execa": "^9.5.2",
+    "execa": "^5.1.1",
     "husky": "^9.1.7",
+    "jest": "^29.7.0",
     "lint-staged": "^15.5.0",
     "oclif": "^4.17.37",
     "prettier": "^3.5.3",
     "shx": "^0.4.0",
+    "ts-jest": "^29.3.1",
     "ts-morph": "^25.0.1",
     "ts-node": "^10.9.2",
     "tslib": "^2.8.1",
